+++ library/src/actions/index.ts
@@ -43,6 +43,7 @@ export * from './maxBytes/index.ts';
 export * from './maxLength/index.ts';
 export * from './maxSize/index.ts';
 export * from './maxValue/index.ts';
+export * from './metadata/index.ts';
 export * from './mimeType/index.ts';
 export * from './minBytes/index.ts';
 export * from './minLength/index.ts';