+++ library/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the library will be documented in this file.
 
 ## vX.X.X (Month DD, YYYY)
 
+- Add `metadata` action to add custom metadata to a schema
 - Add `decimal` action to validate integer and float strings (pull request #823)
 - Rename `NoPipe` type to `SchemaWithoutPipe`
 - Rename `decimal` action to `digits` (pull request #823)